Update patch set 3
Patch Set 3:

I believe we're missing a double-context copy in rd_cost_sbuv_16x16(). Also, what I've noticed is that most RD functions don't actually copy the context result values after the cost_coeffs() is done, so although a[0] is generally set to the new value, for bigger transforms, a[1]-a[n] is not. It may actually make sense to do that update of neighbour values for bigger transforms inside cost_coeffs() also.

Running some tests now both with and without this experiment to see how much it helps and if we can merge this.

Patch-set: 3