Update patch set 2
Patch Set 2:

WRT the swtich statements: Talk to Scott and Yunqing about where they think the boundary is between the asm code and the C code. If we want to switch between c and optimized versions at the vp9_short_iht16x16 level (meaning that the asm deals with the dispatch to different transform types, and that all transform types have the same instruction set), and this file is just the reference then the internal const function pointer tables are fine. But if we want to make vp9_short_iht16x16 a driver that figures out the function pointers to use, and calls the possibly asm versions, then the switch statements are probably better, because the function to call isn't known until runtime beacuse of RTCD.

BTW: if you want a *great* cleanup opportunity, remove all this short pitch nonsense. It's always been crazy that we take the number of bytes instead of the number of elements in these functions.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0