Update patch set 1
Patch Set 1:

Yes, you need to specify struct everywhere -- I consider that a feature. I *want* to know what sort of compound type I'm dealing with, particularly for structs vs unions, but also so that it's obvious that they may be expensive to copy, etc. The extra few characters to type are a small price to pay, imho.

making it struct MV32 is fine (should probably make the MV a named struct too) but I'd prefer we not typedef it and work towards getting rid of the unnecessary typedefs over time. I'd prefer to change the existing references of 'MV' to 'struct MV' and 'struct mv32' to 'struct MV32' than change just the references of 'struct mv32' to 'MV32', despite there being more of them in the former case.

I won't block this patch over the typedef, so do it if you want to make them both consistent, but we need to come to a consensus at some point about which style we're going to use for these things.

Patch-set: 1