Update patch set 1
Patch Set 1:

Also, could you update the commit message with how exactly you're testing? If you're looking for a way to generate the biggest possible frame, it should be possible to *really* big frames with:

  vpxenc --i420 -w 16384 -h 16384 --rc-max-quantizer=0 -o big.webm /dev/urandom

To prove that nothing breaks. I think the fixed size buffer the encoder writes to is probably broken in this case.

Patch-set: 1