Update patch set 4
Patch Set 4: (3 inline comments)

general: It'd be good to preserve the type information on the key as much as possible. I pointed out a few places as making the length explicit, but maybe it'd be better to make it a new type, typedef uint8_t vp8_secret_t[32] or something, in case you ever wanted to change the length of the key. If you put the definition in vp8dx.h, you can probably make it part of the API, which would help in documenting it.

Other than that, mostly lgtm.

Patch-set: 4
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0