Update patch set 3
Patch Set 3:

Don't be fooled into thinking the current behavior was by design :)

We must do some clamping to prevent arbitrary reads, even if that clamp is never invoked by a well formed bitstream.

Making all the modes consistently allow the whole of the block to lie inside the UMV border is probably the right thing to do. This unfortunately means extending the border to 80 pixels for now, and we should work on doing emulated edges instead after release. For clamping while decoding, I think we probably want to only clamp the reference MV when decoding a NEWMV, and then when we go to do prediction do a secondary but non persistent clamp.

Patch-set: 3