Update patch set 1
Patch Set 1: (28 inline comments)

Despite what I said earlier, I'm now mildly opposed to putting this in in its current state. It seems to unnecessarily change code that's not under the experiment, is creating a ton of work to deduplicate all this code later, and at its core I don't think it's taking the right approach -- it hasn't really established a recursive structure. Rectangular modes aren't special cases, the square case should be treated as a special rectangle, even if it's the most common mode. If decoding a SB32 isn't just a PARTITION mode SB64, something's wrong. When doing one of these split modes, we should just be able to divide the width and/or height by two, and call the same function. This should be straightforward until you get down smaller than 16x16, after which you could still do the same thing, but may require more changes to our core data structures.

I'd rather see this go in as a series of smaller patches that lay the infrastructure right rather than try to fix it later. Set up a recursive structure with one split mode, the 4x4 PARTITION mode, combining the 16/32/64 code. This makes the width/height effectively a variable, which makes it easy to go to rectangular modes, so I'd add the other two split modes second. But I suppose you could do the rectangles before the recursion. The important thing is to do it without duplicating the code. In fact, I bet this whole patch series could be net negative, rather than adding 5000 lines.

Patch-set: 1