Update patch set 1
Patch Set 1:

Can you find out what the underlying error is? I mean, could this be a problem for vpxdec too? Calling perror if the write fails might be a simple way to get the error message out as a debugging aid.. also see if ferror() is set. I'd like to understand what the real issue is.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0