Update patch set 6
Patch Set 6:

Rather than working on this as a series of patches to one change, it'd probably be easier to make it a series of changes. Add the types that set up the sizes/log2 stuff, then start using them one functional unit at a time, rather than one big bang. There'd be less blocking of other work doing it that way.

Patch-set: 6