Update patch set 1
Patch Set 1:

I'd rather see this as vp9_read_signed(bc, int) -- make the magnitude of a parameter, and read the magnitude and sign inside the function. The caller doesn't need to know whether it's encoded as sign magnitude or magnitude sign or twos complement or whatever.

Patch-set: 1
Label: Commit-Queue=0