Update patch set 1
Patch Set 1:

Can we get away from block size specific functions, and let them take either the bsize or separate w/h as an argument like we are for the other sb functions? We can get rid of a lot of code if so. This is an implementation detail that can be handled in the asm layer.

Patch-set: 1
Reviewer: John Koleszar <1000856@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0