Update patch set 4
Patch Set 4:

The breakout code (after my fix) isn't actually identical before/after this patch, but it shows improvements on most clips (+0.1% on derf). I think the problem is that the filter_variance subpixel functions in uv_sse aren't very precise when filers (sharp/lowpass) are used, whereas the replacement code is correct about them. So the overall change is probably for the better, and the change in behaviour expected.

Patch-set: 4