Update patch set 1
Patch Set 1:

mixed feelings about this.. I'd ultimately to see us define some syntactic primitives, and then describe the bitstream in terms of them, and have one file that shows how to decode all the primitives, and another that shows how to encode them. These functions skirt both sides of the line. Ok with me to do this for now, but if it starts to make sense to treat them as primitives, let's put them back.

Patch-set: 1
Label: Commit-Queue=0