Expand UMV border to 96 pixels
Ensures that the full 64 pixel border is available for prediction (need a minimum of
64+INTERP_EXTEND on all sides, and 32+INTERP_EXTEND on UV). Value also must be a
multiple of 32 to keep UV stride alignment. The smaller border was causing the prediction
to read outside the frame, which can cause a mismatch.

TODO: Get rid of this explicit border and use edge emulation instead.

Change-Id: I3f68453a088ec0ab4349d0f5cc02b573be06d7c4