Update patch set 1
Patch Set 1:

This does seem to effectively eliminate the selection of the last frame's MV, as the results seem basically neutral between adding and removing that candidate.

I've got two experiments in progress with this. The first is like this patch but using -7 instead of -9 for the last steps of the sb64 neighbors. That one is -0.015 on derf300, +0.149 on YT. The version in this patch is neutral on derf. YT pending. stdhd250 pending on both versions.

There may be better candidates than the ones I've chosen here, this was the first thing I tried. It does seem to show good promise for being able to eliminate the last frame MV, if the results hold up.

Patch-set: 1
Reviewer: John Koleszar <1000856@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0