Update patch set 2
Patch Set 2: (2 inline comments)

So, I don't have much of an opinion on special probabilities for 0/1 w/o vs. w/ skip nodes as separate tables each, but does it make sense to judge them on their own behalf without putting them in a rebalancing patch? That is, maybe we should first commit the rebalancing itself, and then in a separate patch (if it is beneficial and the hw guys are ok with it) add separate tables for yes-skipnode vs. no-skipnode tables.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0