Update patch set 3
Patch Set 3:

mbs_to_right_edge is tile-aware. The mismatches are possibly because you're using a 16x16 tx block and skipping the second on its right, and then the block next to it uses 32x32 tx, but the entropy contexts for that in the second 16x16 tx block above are not initialized. I think the right thing regardless is to set the output entropy coding context for skipped blocks in all stages of encoder and decoder to zero for context purposes in following tx blocks. The same goes for partitioning. Intra prediction edges should be extended.

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0