Update patch set 11
Patch Set 11: Looks good to me, approved

I think I am OK with the concept. It seems a nice solution, though we will have to work out how we handle things like multi-arf from the encoder side as I am pretty sure it will be pretty broken by this patch with its assumptions about buffer usage in the encoder. As long as these fixed assumptions are not present in the decoder though that is fine.

Lets get it in and tested tonight if we can. If we find a problem we can always pull back but given the scope of this change we don't want any extra re-bases.

Patch-set: 11
Reviewer: Paul Wilkins <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2
Label: Commit-Queue=0
Label: Verified=0