Update patch set 1
Patch Set 1: Do not submit

(1 inline comment)

I think this is basically the same as my earlier patch, though mine had an extra test on distortion improvement which made it less aggressive.

The same problem exists with Jim's current speed 1 change which is broken by this so don't merge either for now.

We need to decide which approach is most promising.

Patch-set: 1
Label: Code-Review=-2
Label: Commit-Queue=0