Update patch set 5
Patch Set 5:

Yes. The multiply of qcoeff[i]*dequant[rc!=0] can be 16bits+sign for 32x32, but since we remove the sign and make it an unsigned multiply, it fits again. Then, we rightshift the absolute value (which is the same as /2 for signed value), making it 15bits, and then reintroduce the sign, so it always fits.

Patch-set: 5
Reviewer: Ronald S. Bultje <1000885@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0