Update patch set 1
Patch Set 1:

Is this only for inter-frames, or also for keyframes? For keyframes, I wonder if defaulting to the largest in the same way as we do for inter frames is a good idea (at the very least, it might make sense to distinguish between these two in settings and test each separately to know what good default are for various speed settings).

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0