Update patch set 1
Patch Set 1:

From what I had seen before, the subpel filter decision is somewhat tricky because a frame level rd-optimal decision may actually degrade the performance for subsequent frames more so than in other cases. So we may choose a certain filter typically a smooth one in a certain frame because it is the best for that frame, but that may impact things in subsequent frames because there isn't enough high frequency information preserved.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0