Update patch set 1
Patch Set 1:

Yes, I noticed that especially for the overlay-frame case, it's very typical to think that the smooth filter is the best, whereas it actually is the worst. I haven't figured out why, so had to explicitly disable that option to get it to run positive.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0