Update patch set 1
Patch Set 1:

Can you check whether this gives identical output when scaling is enabled? Thinking about this some more, 90% of struct scale_factors will be the same, but in particular {x,y}_offset_q4 will be different between Y and UV. This should be ok, because we recalculate the scaling offsets before we start scaling each block, and the offsets are only used for that final step so overwriting that variable shouldn't matter, but should probably verify that this doesn't change anything unexpectedly.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0