Update patch set 1
Patch Set 1:

This issue was brought up by Xun Cai (xuncai@google.com). It doesn't change anything in the current 3 reference frames setting. But seems could be an issue if we allow 8 frames in the buffer and select 3 of which for inter-frame prediction(?)

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0