Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Having a stable Q for normal P frames does tend to give better metrics results but makes it hard to be sure of hitting the target rate. Part of the reason for this is that if rate control is applied strictly it is often the case that some sections get starved of bits or are given too many just to hit the rate target.

I am not sure about making this the default as "one shot mode" will not give accurate rate control unless we work hard to improve the algorithm that guesses the Q choice in the first pass. Also there may be problematic rate spikes and I am less convinced that the results are actually much better in terms of psycho-visuals. As such I do not think this mode will be ideal for YT.

The main rationale for using it in testing was to have a mode that is closer in behavior to the fixed Q models used in the HEVC reference code.

Rather than making it the default how about exposing a command line parameter to turn this mode on rather than it being a compile time choice?

Patch-set: 1
Label: Code-Review=-1
Label: Commit-Queue=0