Update patch set 1
Patch Set 1:

I think this is a good catch. The speed thresh feature is now redundant and was clashing with the adaptive rd threshold feature to create a doubled up effect. The former increasing the base rd thresholds based on speed and the later increasing the maximum adaptive multiplier of this. As such I think the speed threshold feature should simply be removed at this juncture as it gives a poor quality speed trade off. (Perhaps you can change this patch to remove the feature completely not just comment out or alternatively I can submit a follow on patch with this and other changes).

I am less convinced by the change regarding square partition feature. In my initial tests the quality differential seems much smaller (say 1-2%) but there is still a substantial speed impact (on bus and akiyo local tests about 20% slower). I am running full tests on cases with one or both changes to get a clearer idea of the quality hit of each.

Even so this is clearly a better trade off for speed 2.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0