Update patch set 2
Patch Set 2:

Fyi : By uploading a change to the commit message I became the sort of committer of record, and this will allow us to test that nothing you did broke compile or build on any other system.  Obviously it also gives you style warnings. :)    

It does not however run any tests on the mips device as we have yet to tie that in to any automated systems.  For now that will have to remain a manual process ..

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0