Update patch set 2
Patch Set 2:

Please take a look now that M31 will be moving to beta soon.

It's a bit hard for me to distinguish what's important from what's not in the changelog (and that's why it's so verbose compared to previous releases - I consider that undesirable), so help would be welcome to remove most entries that are not "important enough" to be in the changelog.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0