Update patch set 8
Patch Set 8:

This patch is ready for code review.

This is a structural change that deprecates the use of RD_SPLITMV, RD_I4X4_PRED, etc. It does not change speed 0 performance.

Will factorize the common codes of the two rd loop, and optimize the operational flow in the next few patches.

Patch-set: 8
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0