Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

As you know, no-prec-div option enables optimizations that give slightly less precise results than full IEEE division. If we are ok with this for all float divides in VP9, we may wish to throw the -mSSE2 or -mSSE3 options as well because then the compiler will opportunistically change floating-point division computations into multiplication by the reciprocal of the denominator. I.E., A/B is computed as A * (1/B) to improve the speed of the computation.

Patch-set: 2
Label: Code-Review=+1
Label: Commit-Queue=0
Label: SUBM=+1