Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

It seems after this the only use of is_inter_mode is within the encoder, so it might be nice to localize that function if it's the only thing that needs that test.

Jingning, any comments on this one?

Patch-set: 2
Label: Code-Review=+1