Update patch set 3
Patch Set 3:

Not sure if I understood why we want to handle them separately? The use of these two contexts are defined as part of the bit-stream, so we want to make sure that their behaviors are identical at both encoder and decoder ends.

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0