Update patch set 2
Patch Set 2:

The intention is not to change the definition of the bit-stream format, but rather to change the behavior of decoder when the input are invalid input bit-stream. This is because that "crash" is not an acceptable behavior for any application  that makes use of the library.

Patch-set: 2
Label: Commit-Queue=0