Update patch set 3
Patch Set 3:

The issue is that the bots Jenkins runs on do not support AVX2. For running libvpx, this is not an issue because we detect at runtime. However, the unit tests run regardless of what cpu features are available. I'm working to disable AVX2 testing on the bots (since they don't have the hardware to run it) for the time being.

Patch-set: 3