Update patch set 5
Patch Set 5:

The style error fix should be amended to this patch instead of committing as a new patch. Here are the steps:
1. Checkout current patch (patch set 5).
2. Fix the warnings.
3. run
$git commit -a --amend
This will amend the current patch "Fix previous SSSE3 bug in the intrinsic version of convolve function", and not create a new one.
Use git log to double check this.
4. push the patch back to Gerrit.

Patch-set: 5