Update patch set 1
Patch Set 1:

This line was intentionally removed in the previous patch to ensure that all frame types are set only in the vp9_set_second_pass_params() function before the actual encoding. I believe it is not needed, and if needed the code to set this should go in the vp9_set_second_pass_params() function by checking the last_frame_type function.

Patch-set: 1