Update patch set 19
Patch Set 19: Abandoned

For reference the CL is here: 
https://gerrit.chromium.org/gerrit/#/c/67968/

This patch as it stands is around 15% slower on cif clips with backwards updates and around 4% slower on 1080p clips without them. Its only been tested on 64 bit linux,   I think there is still a bug that makes it not work on others.  

The code reads the first 1, 2 or 3 bits of each token by checking a table that holds start values to compare against if the 3 byte value is less than the first eob is decoded, else if less than the second a 0 is decoded, if less than the third a 1 is decoded otherwise we read the rest of the bits the normal way to determine which token greater than 1 we have.   After this determination is made the bool decoder state is restored from the table ( range and counter).   The decode operation ( outside of the table read) to make the first call is somewhere in the neighborhood of 1/3 the number of operations to read the first 3 1, 2 or 3 bits than reading in the normal way.

However this code incurs some additional costs that account for the slowdown: 

1) The table lookup -> is large: The table itself holds as many entries for a token as their are contexts.   For reference that's : TX_SIZES(4) *  BLOCK_TYPES(2) * REF_TYPES(2) * COEF_BANDS(6) * PREV_COEF_CONTEXTS(6) * LAST_IS_ZERO(2)  * RANGE(128)  * SIZEOF_STRUCTURE (20)  and we jump all over it.. So there's cache miss opportunity.

2) The bool decoder value is probably refilled more frequently ( to be safe)

3) You must fill the entries of the table.   The filling operation in the code was optimized to remove excess work but still remains in the neighborhood of 15 times as expensive as a single 3 token bool decode.  This is in part because because I always completely fill a table entry with all possible choices where as the decoder only takes the one path it needs.  Unfortunately, on the clips tested reuse of a lazy filled table was only 2-5 times as likely as a new fill ( that means in affect that a large number of entries are only visitted one time and are therefore not worth the cost. I also tested prefill operation.  It should also be noted that using a lazy fill incurs an if and a lookup into the structure,  and prefill means you fill a whole lot of entries that never get used.   

4) In the case of a lazy fill you must empty the cache on every frame that switches contexts,  or updates the entropy in either backwards or forwards capacity.   Given the size of the structure even the fact that we only zero out one byte costs a good amount.     

A few things I tried along the way: 

- Altered number of bits read,  ( 2 or 3) 

- Tried only reading first token not a whole coefficient band

- Prefilling sections of the table,  all the table or none of the table.  

There are a few things I've come up with that we could still explore in this code despite my abandonment: 

A) Filling all 128 range entries can be done faster and with a smaller table.  Multiple entries in the same table above reuse the same values.   That this would be so is obvious given a cursory examination of the basic operation :    (range * probability) >>8.   Ranges of 128 and 129 will give exactly the same results if probability is 128.    For low probabilities multiple additional ranges can give the same values for probability 4 for instance 128 - 191 all give the same resultant range.  This fact can be used to save operations in the table setup.  Adding an additional lookup from the table to a pointer that holds the whole 20 byte structure to save space in the table...

B) The table setup has no dependencies and could be done in a completely separate thread as soon as the updates are known.  

C) We could limit the usage however we like - for instance we only use on y 4x4 blocks...

D) We could make the bool decoder keep track of the start value we need to set up the table and lazy fill only the specific range we know ( to speed up the table setup) and still get a value out.  

E) We could try keeping an LRU hash table with just probabilities as the lookup mapped to single 256 bit space to keep the table size down.  ( but creation of hash plus non frequent use of cached entries likely mean this won't work great)...

I don't think any of these get around the fact that the entries in the table aren't re visited enough at even the relatively high bitrates,  I was testing.    1.2 megabits for cif,  and 10 megabits for 1080p..

If the bitrates are significantly higher and we aren't doing backwards updates its conceivable that the equation would change.

Patch-set: 19